Simplified dependencies
