+++ .env.example
@@ -1,10 +1,4 @@
-OUTPUT_DIR=target/celi_output
 OPENAI_API_KEY=sk-XXXX
-DB_URL=mongodb://localhost:27017/
-EXTERNAL_DB=True
-NO_MONITOR=True
-NO_CACHE=True
+NO_CACHE=False
 JOB_DESCRIPTION=celi_framework.examples.wikipedia.job_description.job_description
 TOOL_CONFIG_JSON=celi_framework/examples/wikipedia/example_config.json
-PARSER_MODEL_CLASS=llm_core.parsers.OpenAIParser
-PARSER_MODEL_NAME=gpt-3.5-turbo-16k
