+++ celi_framework/examples/wikipedia/tools.py
@@ -12,7 +12,8 @@ from pydantic import BaseModel
 from celi_framework.core.celi_update_callback import CELIUpdateCallback
 from celi_framework.core.job_description import BaseDocToolImplementations
 from celi_framework.examples.wikipedia.index import get_wikipedia_index
-from celi_framework.utils.codex import MongoDBUtilitySingleton
+from celi_framework.experimental.codex import MongoDBUtilitySingleton
+from celi_framework.utils.llm_cache import get_celi_llm_cache
 from celi_framework.utils.utils import format_toc, get_section_context_as_text
 
 logger = logging.getLogger(__name__)
@@ -271,14 +272,14 @@ class WikipediaToolImplementations(BaseDocToolImplementations):
         ]
         return top_nodes
 
-    def ask_question_about_target(self, prompt: str):
+    async def ask_question_about_target(self, prompt: str):
         """
         Ask a natural language question about the target subject and get a response.  This can be used to fill in gaps that are not present in the references.
 
         Args:
             prompt (str): The question to ask.
         """
-        result = self._get_codex().check_llm_cache(
+        result = await get_celi_llm_cache().check_llm_cache(
             target_url=self.target_url, prompt=prompt
         )
         if result:
@@ -289,7 +290,7 @@ class WikipediaToolImplementations(BaseDocToolImplementations):
             logger.debug("Caching LLM response")
             result = self.target_query_engine.query(prompt)
             result_dict = self._response_to_dict(result)
-            self._get_codex().cache_llm_response(
+            await self._get_codex().cache_llm_response(
                 response={"response": result_dict},
                 target_url=self.target_url,
                 prompt=prompt,
