+++ celi_framework/utils/llms.py
@@ -25,6 +25,7 @@ Features and Functionalities:
 
 import asyncio
 import functools
+import logging
 import os
 import re
 import time
@@ -35,13 +36,14 @@ from openai.types.chat import ChatCompletion
 from pydantic import BaseModel
 from requests import HTTPError
 
-from celi_framework.utils.codex import MongoDBUtilitySingleton
-from celi_framework.utils.exceptions import ContextLengthExceededException
+from celi_framework.utils.llm_cache import get_celi_llm_cache
 from celi_framework.utils.log import app_logger
 from celi_framework.utils.token_counters import (
     token_counter_decorator_ask_split,
     token_counter_decorator_quick_ask,
 )
+
+logger = logging.getLogger(__name__)
 
 
 # Initialize the OpenAI client, using the OPENAI_API_KEY environment variable.
@@ -70,7 +72,6 @@ async def ask_split(
     wait_between_retries=2,
     temperature=0.0,
     timeout: Optional[int] = 120,
-    codex: Optional[MongoDBUtilitySingleton] = None,
     tool_descriptions: Optional[List[ToolDescription]] = None,
     model_url: Optional[str] = None,
     json_mode: bool = False,
@@ -89,7 +90,6 @@ async def ask_split(
                 app_logger.error(f"Attempt {err_cnt + 1}:", extra={"color": "red"})
             if model_url is None:
                 chat_completion = await cached_chat_completion(
-                    codex=codex,
                     base_url=model_url,
                     messages=[
                         {"role": "system", "content": system_message},
@@ -113,7 +113,6 @@ async def ask_split(
                 )
             else:
                 chat_completion = await cached_chat_completion(
-                    codex=codex,
                     base_url=model_url,
                     messages=[
                         {"role": "system", "content": system_message},
@@ -168,7 +167,6 @@ def quick_ask(
     timeout=90,
     time_increase=30,
     model_url: Optional[str] = None,
-    codex: Optional[MongoDBUtilitySingleton] = None,
 ):
     """
     Simplified version of ask_split for quickly sending prompts to the OpenAI API, with error handling and retries.
@@ -213,7 +211,6 @@ def quick_ask(
 
             chat_completion = asyncio.run(
                 cached_chat_completion(
-                    codex=codex,
                     base_url=model_url,
                     messages=assemble_chat_messages(prompt),
                     model=model_name,
@@ -284,22 +281,23 @@ def assemble_chat_messages(prompt: str | List[Tuple[str, str] | Dict[str, str]])
         return [format_message(msg) for msg in prompt]
 
 
-async def cached_chat_completion(
-    codex: Optional[MongoDBUtilitySingleton], base_url: Optional[str], **kwargs
-) -> ChatCompletion:
-    if codex:
+async def cached_chat_completion(base_url: Optional[str], **kwargs) -> ChatCompletion:
+    cache = get_celi_llm_cache()
+    if cache:
         url_dict = {} if base_url is None else {"base_url": base_url}
         cache_args = {**url_dict, **kwargs}
-        ret = codex.check_llm_cache(**cache_args)
+        ret = await cache.check_llm_cache(**cache_args)
         if ret:
             app_logger.debug("Using cached LLM response")
+            if isinstance(ret, str):
+                logger.info(ret)
             return ChatCompletion.model_validate(ret["completion"])
         else:
             app_logger.debug("Caching LLM response")
             result = await get_openai_client(
                 base_url=base_url,
             ).chat.completions.create(**kwargs)
-            codex.cache_llm_response(
+            await cache.cache_llm_response(
                 response={"completion": result.dict()}, **cache_args
             )
             return result
@@ -307,3 +305,11 @@ async def cached_chat_completion(
         return await get_openai_client(base_url=base_url).chat.completions.create(
             **kwargs
         )
+
+
+class ContextLengthExceededException(Exception):
+    """Exception raised when the input exceeds the model's maximum context length."""
+
+    def __init__(self, message="Context length exceeded the model's maximum limit."):
+        self.message = message
+        super().__init__(self.message)
