+++ pyproject.toml
@@ -13,25 +13,33 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Typing :: Typed",
 ]
-include = ["celi_framework/examples/wikipedia/example_config.json"]
+include = ["celi_framework/utils/seed_llm_cache.db", "celi_framework/examples/wikipedia/example_config.json"]
 
 [tool.poetry.dependencies]
 python = "^3.10"
-bert-score = ">=0.3.13"
-evaluate = ">=0.4.1"
-py-llm-core = ">=2.8.2"
-pymongo = ">=4.6.2"
 python-dotenv = ">=1.0.0"
-requests-cache = ">=1.2.0"
-llama-index = ">=0.10.23"
-llama-index-vector-stores-chroma = ">=0.1.6"
-scikit-learn = ">=1.4.2"
-pymupdf = ">=1.24.1"
-python-docx = ">=1.1.0"
-jupyter = ">=1.0.0"
-timeout-decorator = ">=0.5.0"
-opentelemetry-proto = ">=1.24.0"
-python-levenshtein = "^0.25.1"
+python-levenshtein = ">=0.25.1"
+aiosqlite = ">=0.20.0"
+numpy = ">=1.22.4"
+openai = ">=1.16.0"
+pandas = ">=2.0.3"
+pydantic = ">=2.0.1"
+requests = ">=2.26.0"
+llama-index = { version = ">=0.10.23", optional = true }
+llama-index-vector-stores-chroma = { version = ">=0.1.6", optional = true }
+py-llm-core = { version = ">=2.8.2", optional = true }
+opentelemetry-proto = { version = ">=1.24.0", optional = true }
+pymongo = { version = ">=4.6.2", optional = true }
+bert-score = { version = ">=0.3.13", optional = true }
+evaluate = { version = ">=0.4.1", optional = true }
+requests-cache = { version = ">=1.2.0", optional = true }
+scikit-learn = { version = ">=1.4.2", optional = true }
+pymupdf = { version = ">=1.24.1", optional = true }
+python-docx = { version = ">=1.1.0", optional = true }
+
+[tool.poetry.extras]
+experimental = ["py-llm-core", "opentelemetry-proto", "pymongo", "bert-score", "evaluate", "requests-cache", "scikit-learn", "pymupdf", "python-docx"]
+wikipedia = ["llama-index", "llama-index-vector-stores-chroma"]
 
 [tool.poetry.group.test.dependencies]
 pytest = ">=7.4.4"
@@ -39,6 +47,7 @@ pytest-asyncio = ">=0.23.3"
 
 [tool.poetry.group.dev.dependencies]
 ipykernel = ">=6.29.3"
+jupyter = ">=1.0.0"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.2.6"
